reorganize specific_area_patches
hash changes are differences in instance ID/dependency order due to some patch functions being called in a different order than before
