+++ src/open_prime_rando/echoes/specific_area_patches/__init__.py
@@ -1,3 +1,17 @@
-from open_prime_rando.echoes.specific_area_patches import rebalance_patches, required_fixes, version_differences
+from open_prime_rando.echoes.specific_area_patches import (
+    dynamic_loading,
+    front_end,
+    quality_of_life,
+    rebalance_patches,
+    required_fixes,
+    version_differences,
+)
 
-__all__ = ["rebalance_patches", "required_fixes", "version_differences"]
+__all__ = [
+    "dynamic_loading",
+    "front_end",
+    "quality_of_life",
+    "rebalance_patches",
+    "required_fixes",
+    "version_differences",
+]
