+++ tests/test_files/ntsc_hashes.json
@@ -7230,7 +7230,7 @@
             "MAPA_f3e204aa": "e55ab5b5b378cbaed4748b31c4d9e0dc32f79e6749d051f9a9098cb93f315b38",
             "MAPA_fff2fb36": "4a98f20a4b520722074f46b8906b7a62e5bd52d781fc310b3b88ac8c69c388b0",
             "MAPW_4198fdbe": "f844e4269c35a0c8f6b96a99f17591777a95bb33437ecee00c7be0e568e356d4",
-            "MLVL_3bfa3eff": "0decfcb20b2de4cf9163c75359aef50e1d9fb6ad551a89168b739377216bb149",
+            "MLVL_3bfa3eff": "7e4b9e82ef2c2e2206782826a0c8a1f939e488d12a013c6791fb3f87987c4cc9",
             "MREA_0405eba4": "6d842ac118a1cc5eecee504830dcee8e57c761a3a0b48092a921f4e8b78750c9",
             "MREA_131c3388": "3f75a0c8f19eb283397bd5c5cda3185d40d36a06ba9de744bdd28397f3e9327c",
             "MREA_1461dc2f": "01b477c4ae74ba56b92d1cdde644fac7900694ab0489714518090ff186cee7a1",
@@ -7249,7 +7249,7 @@
             "MREA_50b372a5": "6f45cca29961983fa8c57d95e38cf855de745a59378fab6b3208a4556ac335b7",
             "MREA_531079ba": "cf25f7028cc4d39e18ddf3eb9f43dc09858ec9c03959099a7cd5dca53b3fb631",
             "MREA_55ee6dc3": "f8c4bf97bc73ab5d658202f3a49a7f5062fd8ba3e94751b7e400531d258ec941",
-            "MREA_57757ad5": "c67eda4b9dda725d6abb18e8d053949905baa1e6f52c766ec24426377ce4fc86",
+            "MREA_57757ad5": "368eda598bee5b2f8029387d3c83b7202b72c5ca5ffba064e836b0045177ab50",
             "MREA_5b92d146": "bb0a18c827c756d89a512c5761613ec32775b40088328516bbc92ed1a6fde0b7",
             "MREA_62ac8ac4": "a77b2eb38456c9f19b78a38911f358a41d9d38dfe84de0925956f61804bb9a65",
             "MREA_62b0d67d": "4c06cca6669907b6f4205918f13a52ff2305f615be09d01acb0d7e6b840a2503",
@@ -17119,7 +17119,7 @@
             "MREA_bf7f572c": "6a36d8ada969ae01f5298a7861d983cfcfc8319b0d401e36610f540a377235ba",
             "MREA_c5c7c9d3": "8a21a6e17d739d81b424b4e99288ab55ce6ff6821d60aa6e2f6dfd914109790d",
             "MREA_c68b5b51": "b6758b592368610a0690356ae884c0231064e273bc69cf368f8e6eb1ac7fbb44",
-            "MREA_ccd9f39e": "ec33c6f80852b26e398f2b12861c93e78a0411612e4fa315f72e8c5c68d260b5",
+            "MREA_ccd9f39e": "e8a771270073cb1cc5afa5719adc79d324ce3be47ca15dd199ad0c169efbeed8",
             "MREA_d73ecd68": "c58efa969e6c4c603a9d4fa4f9c2af9b5f8a3b5871869ed2d3f5ef1cf75af83a",
             "MREA_d7a15431": "6bcbbbaadaf1d927ba56c511ff032c9fabe4dde8f0d5b879652810a326e0dc45",
             "MREA_d92890c3": "0ed83c34a7e59fd333df0de640e4a579449ba4ea14fd5a568d4024aa25e8decd",
@@ -24133,7 +24133,7 @@
             "MREA_0caaeec6": "f251fad32a138c3e3d72acc59c52239631a3e4662cff6b6887743486fc3b51bd",
             "MREA_11c4526b": "eaf45e98aef039e12c3fcc265f8b74f69569d49aaa0561991c4819fc478ed0f5",
             "MREA_13254ec6": "9aa2401ba3e7786698579b40843e5ad7c8c33365be4f7521c0602762944e13b6",
-            "MREA_133bf5b8": "8febadf5f5157f55641446a4b6532d4efdf7402ecf5d43cdf2d46b5e2bd2483b",
+            "MREA_133bf5b8": "01812c68b45b447fe5338a4fe9cc47121e8f0febab3dd52533f5fe89b2431964",
             "MREA_14e2c3a6": "1a874624fbfe5bf8a4fefc4fa3f9679bfa556d02028245a8bce3b0dcfd7d13be",
             "MREA_27d2c898": "f20ea8be2d8f7ea56a306bd23a6ef0dab46c1b19f4cb7fcf60fb3bfde60e3dd5",
             "MREA_282b587f": "5d7ef331b550c2d1fe904030727e37b7aa147c4a7233a7f82af46d6e8e4670b5",
