+++ tests/test_files/pal_hashes.json
@@ -7234,7 +7234,7 @@
             "MAPA_f3e204aa": "e55ab5b5b378cbaed4748b31c4d9e0dc32f79e6749d051f9a9098cb93f315b38",
             "MAPA_fff2fb36": "4a98f20a4b520722074f46b8906b7a62e5bd52d781fc310b3b88ac8c69c388b0",
             "MAPW_4198fdbe": "f844e4269c35a0c8f6b96a99f17591777a95bb33437ecee00c7be0e568e356d4",
-            "MLVL_3bfa3eff": "2ee7912cc97da08ef94856329a06d9b86f17be8c284997673d2d4dfab0d757f2",
+            "MLVL_3bfa3eff": "2f1a2555f0f460fdf5d840f856d083fda9113f8ad8dabafa6163cd370f6c92ca",
             "MREA_0405eba4": "6d842ac118a1cc5eecee504830dcee8e57c761a3a0b48092a921f4e8b78750c9",
             "MREA_131c3388": "3f75a0c8f19eb283397bd5c5cda3185d40d36a06ba9de744bdd28397f3e9327c",
             "MREA_1461dc2f": "01b477c4ae74ba56b92d1cdde644fac7900694ab0489714518090ff186cee7a1",
@@ -7253,7 +7253,7 @@
             "MREA_50b372a5": "cd8aedb8418c6b6b562f9cfc05b35fa73b4426574674601ca2e5d71eefed6f15",
             "MREA_531079ba": "cf25f7028cc4d39e18ddf3eb9f43dc09858ec9c03959099a7cd5dca53b3fb631",
             "MREA_55ee6dc3": "f8c4bf97bc73ab5d658202f3a49a7f5062fd8ba3e94751b7e400531d258ec941",
-            "MREA_57757ad5": "98b052a0212e66b0552d54bb9b99ddb4627211c804251130dd81e9b8b29bc3fd",
+            "MREA_57757ad5": "8d7cc35ab086ba6df9febe95eecaaf0f6c47fff4f6426149fcb3ad35b8c7b170",
             "MREA_5b92d146": "bb0a18c827c756d89a512c5761613ec32775b40088328516bbc92ed1a6fde0b7",
             "MREA_62ac8ac4": "1fc160db7129009c0abe8fd4bf37be5e42a7e4e1c5acd8522f79391eb3d1cc19",
             "MREA_62b0d67d": "4c06cca6669907b6f4205918f13a52ff2305f615be09d01acb0d7e6b840a2503",
@@ -17125,7 +17125,7 @@
             "MREA_bf7f572c": "c26c909e20aab9f3c885de736bb1788d35faf537b37fa4abc7a59ebe1c85ea62",
             "MREA_c5c7c9d3": "20ab334f213df025e69ae4f2932018c04306fdd51b4bc6b3fe48da6c62534f46",
             "MREA_c68b5b51": "b6758b592368610a0690356ae884c0231064e273bc69cf368f8e6eb1ac7fbb44",
-            "MREA_ccd9f39e": "ec33c6f80852b26e398f2b12861c93e78a0411612e4fa315f72e8c5c68d260b5",
+            "MREA_ccd9f39e": "e8a771270073cb1cc5afa5719adc79d324ce3be47ca15dd199ad0c169efbeed8",
             "MREA_d73ecd68": "c58efa969e6c4c603a9d4fa4f9c2af9b5f8a3b5871869ed2d3f5ef1cf75af83a",
             "MREA_d7a15431": "6bcbbbaadaf1d927ba56c511ff032c9fabe4dde8f0d5b879652810a326e0dc45",
             "MREA_d92890c3": "0ed83c34a7e59fd333df0de640e4a579449ba4ea14fd5a568d4024aa25e8decd",
@@ -24151,7 +24151,7 @@
             "MREA_0caaeec6": "6f9e5324d67025036103176ea8d6195601da1e42dbbed04cb72a5f990cdfe6e8",
             "MREA_11c4526b": "e824f58bbaf21c51041368539aef5bc34ecee18ddcb428ba301e4b7ef9afd56f",
             "MREA_13254ec6": "9aa2401ba3e7786698579b40843e5ad7c8c33365be4f7521c0602762944e13b6",
-            "MREA_133bf5b8": "7379fe5e4e8b1dc96fb37d1218a2360a7ea947baa7321660cc0adee4c566946b",
+            "MREA_133bf5b8": "3bbf4fa755b0c529039b18f1dd6edf4532b38a0614618064dbbd5fd26f148bf0",
             "MREA_14e2c3a6": "1a874624fbfe5bf8a4fefc4fa3f9679bfa556d02028245a8bce3b0dcfd7d13be",
             "MREA_27d2c898": "3acd56a434dbab817a138f11824ca006f7e32143aafb43aa484b54e302af4948",
             "MREA_282b587f": "7ac63d5cb8397c5f8b7e3c9c69716abb5b2057d62ba9924d7440ad8e6913f325",
